+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Add the end point about Dungeon's creatures.
 - Add the end point about Stronghold's creatures.
 - Add the end point about Fortress's creatures.
+- Add the end point about Conflux's creatures.